requirements.txt
# Core
transformers>=4.43.0
accelerate>=0.33.0
torch>=2.3.0
safetensors>=0.4.3
diffusers>=0.30.0
sentencepiece>=0.2.0
protobuf>=4.25.3

# UI
streamlit>=1.36.0
Pillow>=10.4.0

# Perf
xformers>=0.0.27.post2; sys_platform == 'win32'
einops>=0.8.0

# Misc
huggingface_hub>=0.24.5
numpy>=1.26.4
